+++ src/if.vue
@@ -1,6 +1,13 @@
 <script setup>
   import { ref } from 'vue'
   const awesome = ref(true)
+
+  function toggleAwesome(v) {
+    awesome.value = !v
+}
+  function isAwesome() {
+    return awesome.value
+}
 </script>
 
 <template>
@@ -8,4 +15,7 @@
 
   <h1 v-if="awesome"> Vue is awesome!</h1>
   <h1 v-else> Oh no </h1>
+
+  <h4 v-if="isAwesome">v-if TRUE</h4>
+  <h4 v-show="isAwesome">v-show TRUE</h4> 
 </template>
